undo class
