+++ Midterm_FullStack_Assignment1_RegisterLogin/Controllers/AccountController.cs
@@ -99,4 +99,33 @@ public class AccountController : Controller
         var passwordRegex = new Regex(@"^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^\da-zA-Z]).{8,}$");
         return passwordRegex.IsMatch(password);
     }
+
+    public string HashPassword(string password)
+    {
+        // Hash the password using PBKDF2 with a random salt
+        byte[] salt;
+        new RNGCryptoServiceProvider().GetBytes(salt = new byte[16]);
+        var pbkdf2 = new Rfc2898DeriveBytes(password, salt, 10000);
+        byte[] hash = pbkdf2.GetBytes(20);
+        byte[] hashBytes = new byte[36];
+        Array.Copy(salt, 0, hashBytes, 0, 16);
+        Array.Copy(hash, 0, hashBytes, 16, 20);
+        return Convert.ToBase64String(hashBytes);
+    }
+
+    public bool VerifyPassword(string enteredPassword, string hashedPassword)
+    {
+        // Verify the entered password against the hashed password
+        byte[] hashBytes = Convert.FromBase64String(hashedPassword);
+        byte[] salt = new byte[16];
+        Array.Copy(hashBytes, 0, salt, 0, 16);
+        var pbkdf2 = new Rfc2898DeriveBytes(enteredPassword, salt, 10000);
+        byte[] hash = pbkdf2.GetBytes(20);
+        for (int i = 0; i < 20; i++)
+        {
+            if (hashBytes[i + 16] != hash[i])
+                return false;
+        }
+        return true;
+    }
 }
